deployment/requirements.txt
streamlit
streamlit-extras
Pillow
tensorflow==2.15
numpy
python-dotenv
pymongo
boto3
opencv-python-headless
imutils
